refactor(providers): remove redundant type facades
