+++ lambdas/libs/runner-providers/aws/ec2/src/webhook/dynamic-labels-policy.ts
@@ -1,4 +1,4 @@
-import type { AwsDynamicLabelsPolicy, AwsDynamicLabelsValueRule } from './types';
+import type { AwsDynamicLabelsPolicy, AwsDynamicLabelsValueRule } from '../../../../contracts';
 
 export type Ec2DynamicLabelsValueRule = AwsDynamicLabelsValueRule;
