--- lambdas/libs/runner-providers/aws/ec2/src/webhook/types.ts
@@ -1,7 +1,0 @@
-export type {
-  AwsDynamicLabelsPolicy,
-  AwsDynamicLabelsValueRule,
-  DynamicLabelDispatchTarget,
-  DynamicLabelProvider,
-  RunnerMatcherConfig as Ec2RunnerMatcherConfig,
-} from '../../../../contracts';
